chore: clean up index to look cleaner
