+++ packages/api/src/middleware/auth.middleware.ts
@@ -1,55 +1,71 @@
-import type { Context } from 'hono';
+import { Context } from 'hono';
+import { cors } from 'hono/cors';
+import { AppContext } from '../utils/context';
 import { env } from 'hono/adapter';
-import { verifyRequestOrigin } from 'lucia';
 
-import { readBearerToken } from '@repo/api/utils/auth';
-import type { AppContext } from '@repo/api/utils/context';
-import { validateSessionToken } from '@repo/api/utils/sessions';
-
-export const AuthMiddleware = async (
+/**
+ * Handle session from auth middleware.
+ *
+ * This middleware will fetch the session from auth API for every request and
+ * store it in the context. If the session is not found, it will set the user and
+ * session to null in the context.
+ *
+ * @param c - The context object
+ * @param next - The next function to run
+ */
+export async function handleSessionMiddleware(
   c: Context<AppContext>,
   next: () => Promise<void>,
-) => {
-  if (c.req.path.startsWith('/auth')) {
+) {
+  const auth = c.get('auth'); // Retrieve auth from context
+  if (!auth) {
+    console.error('Auth is not initialized');
     return next();
   }
-  const originHeader = c.req.header('Origin') ?? c.req.header('origin');
-  const hostHeader = c.req.header('Host') ?? c.req.header('X-Forwarded-Host');
 
-  if (
-    (!originHeader ||
-      !hostHeader ||
-      !verifyRequestOrigin(originHeader, [hostHeader, env(c).WEB_DOMAIN])) &&
-    env(c).WORKER_ENV === 'production' &&
-    c.req.method !== 'GET'
-  ) {
-    return new Response(null, {
-      status: 403,
-    });
-  }
+  const session = await auth.api.getSession({ headers: c.req.raw.headers });
 
-  const authorizationHeader = c.req.header('Authorization');
-  const bearerSessionId = readBearerToken(authorizationHeader ?? '');
-  const sessionId = bearerSessionId;
-
-  if (!sessionId) {
-    return new Response('Unauthorized', { status: 401 });
-  }
-
-  const { session, user } = await validateSessionToken(sessionId, c);
   if (!session) {
-    return new Response('Unauthorized', { status: 401 });
+    c.set('user', null);
+    c.set('session', null);
+    return next();
   }
-  // LATER: check out lucia auth and see what session?fresh is
-  // - can add another part of fresh in the schema
-  // if (session?.fresh) {
-  // 	const sessionCookie = await createSession(user.id, session.id, c);
-  // 	const serializedCookie = `sessionId=${
-  // 		sessionCookie.id
-  // 	}; Path=/; HttpOnly; Secure; SameSite=Strict; Expires=${sessionCookie.expiresAt.toUTCString()}`;
-  // 	c.header("Set-Cookie", serializedCookie);
-  // }
+
+  const user = {
+    ...session.user,
+    image: session.user.image ?? null,
+  };
+
+  const sessionData = {
+    ...session.session,
+    ipAddress: session.session.ipAddress ?? null,
+    userAgent: session.session.userAgent ?? null,
+  };
+
   c.set('user', user);
-  c.set('session', session);
-  await next();
-};
+  c.set('session', sessionData);
+  return next();
+}
+
+/**
+ * CORS Middleware for better authentication.
+ *
+ * This middleware configures Cross-Origin Resource Sharing (CORS) settings
+ * for the application. It allows the application to handle requests from
+ * specified origins and supports specific HTTP methods and headers.
+ *
+ * @param origin - The allowed origins for CORS requests
+ * @param allowHeaders - The headers allowed in CORS requests
+ * @param allowMethods - The HTTP methods allowed in CORS requests
+ * @param exposeHeaders - The headers exposed to the client
+ * @param maxAge - The maximum age for the CORS preflight request cache
+ * @param credentials - Whether credentials are supported in CORS requests
+ */
+export const betterAuthCorsMiddleware = (c: Context<AppContext>) => cors({
+  origin: [env(c).WEB_DOMAIN || 'http://localhost:3000'], // Use env var for frontend domain
+  allowHeaders: ['Content-Type', 'Authorization'],
+  allowMethods: ['POST', 'GET', 'OPTIONS'],
+  exposeHeaders: ['Content-Length'],
+  maxAge: 600,
+  credentials: true, // Required for cookies to work cross-origin
+});
