+++ packages/api/src/routers/index.ts
@@ -1,2 +1,2 @@
-export * from '@repo/api/routers/hello';
+export * from '@repo/api/routers/auth';
 export * from '@repo/api/routers/user';
