+++ packages/api/src/routers/user.ts
@@ -2,17 +2,19 @@ import { Hono } from 'hono';
 
 import type { AppContext } from '@repo/api/utils/context';
 
-export const userRouter = new Hono<AppContext>().get('/', (c) => {
-  const user = c.get('user');
-  return c.json(user);
-});
-// .get('/oauth-accounts', async (c) => {
-//   const oauthAccounts = await c.get('db').query.oauthAccountTable.findMany({
-//     where: (u, { eq }) => eq(u.userId, c.get('user')?.id ?? ''),
-//   });
-//   return c.json({
-//     accounts: oauthAccounts.map((oa) => ({
-//       provider: oa.provider,
-//     })),
-//   });
-// });
+export const userRouter = new Hono<AppContext>()
+  .get('/', (c) => {
+    const user = c.get('user');
+    return c.json(user);
+  })
+  .get('/session', async (c) => {
+    const session = c.get('session');
+    const user = c.get('user');
+
+    if (!user || !session) return c.body(null, 401);
+
+    return c.json({
+      session,
+      user,
+    });
+  });
